refactor: move the default PTY registry into registry.ts (file-size cap)
